Manipulate alpha using KeyFrame instead of constraint
By using a KeyAttribute inside a keyframe, you don't need to redefine all the constraints of the description and text since you aren't chaging their constraints but are only chaing their attribute. If you defined it by using a custom attribute inside a constraint, then you would need to include all the constraints as well as that part is reapplied. In our example, that doesn't make sense and just a KeyFrame is enough